+++ vshard_test.go
@@ -1,6 +1,7 @@
 package vshard_router //nolint:revive
 
 import (
+	"sync/atomic"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -31,8 +32,10 @@ func TestRouter_RouterBucketCount(t *testing.T) {
 
 func TestRouter_RouteMapClean(t *testing.T) {
 	r := Router{
-		cfg:      Config{TotalBucketCount: 10},
-		routeMap: make([]*Replicaset, 10),
+		cfg: Config{TotalBucketCount: 10},
+		view: &consistentView{
+			routeMap: make([]atomic.Pointer[Replicaset], 10),
+		},
 	}
 
 	require.NotPanics(t, func() {